requirements.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
mysql-connector-python==8.0.33
openpyxl==3.1.2
pandas==2.0.3
aiohttp==3.8.4
